requirements.txt
amqp==5.0.5
appdirs==1.4.4
argcomplete==1.12.2
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
boto3==1.16.45
botocore==1.19.45
celery==5.0.5
certifi==2020.11.8
cffi==1.14.3
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-admin-interface==0.14.2
django-colorfield==0.3.2
django-cors-headers==3.5.0
django-countries==6.1.3
django-environ==0.4.5
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-npm==1.0.0
django-oauth-toolkit==1.3.3
django-pipeline==2.0.5
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-social-oauth2==1.0.8
drf-yasg==1.20.0
durationpy==0.5
future==0.18.2
gunicorn==20.0.4
hjson==3.0.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
kappa==0.6.0
kombu==5.0.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
parameterized==0.7.4
pathspec==0.8.1
pem==20.1.0
Pillow==8.0.1
pip-tools==5.4.0
placebo==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
qrcode==6.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
segno==1.3.1
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
stripe==2.55.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.55.0
troposphere==2.6.3
twilio==6.51.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
xlrd==2.0.1
zappa==0.52.0
